Change coordinates type to double

+++ XayyatMario/src/main/java/com/mammedbrk/model/component/block/Block.java
@@ -22,7 +22,7 @@ public abstract class Block extends Component {
     public Block() {
     }
 
-    public Block(int x, int y) {
+    public Block(double x, double y) {
         super(x, y);
     }
 }
