+++ XayyatMario/src/main/java/com/mammedbrk/model/component/item/Flower.java
@@ -4,7 +4,7 @@ public class Flower extends Item {
     public Flower() {
     }
 
-    public Flower(int x, int y) {
+    public Flower(double x, double y) {
         super(x, y);
     }
 }
